feat: Completed +,-,x,/ logic
This commit has the calculator fully working with add,
subtract, multiply, and divide.

In addition, one of the operands (+,-,x,/) is pressed
more than once before pressing =, it doesn't do anything.

Also, if an operation is done (like 2+3), then you add
(+ 3) again, it solves the first part (2+3), displays
the answer, then it solves the 2nd part (+ 3). This
works for all operands.

Finally, if the = sign is clicked, solving the calculation,
can continue using operands to future solve the calculation.